+++ libs/poses/include/mrpt/poses/CPose3D.h
@@ -129,14 +129,42 @@ class CPose3D : public CPose<CPose3D, 6>,
 	/** Default constructor, with all the coordinates set to zero. */
 	CPose3D();
 
-	/** Constructor with initilization of the pose; (remember that angles are
-	 * always given in radians!)  */
+	/** Constructor with initilization of the pose, translation (x,y,z) in
+	 * meters, (yaw,pitch,roll) angles in radians.
+	 *
+	 * \sa FromXYZYawPitchRoll()
+	 */
 	CPose3D(
 		const double x, const double y, const double z, const double yaw = 0,
 		const double pitch = 0, const double roll = 0);
 
 	/** Returns the identity transformation */
 	static CPose3D Identity() { return CPose3D(); }
+
+	/** Builds a pose from a translation (x,y,z) in
+	 * meters and (yaw,pitch,roll) angles in radians. \note (New in MRPT 2.1.8)
+	 */
+	static CPose3D FromXYZYawPitchRoll(
+		double x, double y, double z, double yaw, double pitch, double roll)
+	{
+		return CPose3D(x, y, z, yaw, pitch, roll);
+	}
+
+	/** Builds a pose with a null translation and (yaw,pitch,roll) angles in
+	 * radians. \note (New in MRPT 2.1.8)
+	 */
+	static CPose3D FromYawPitchRoll(double yaw, double pitch, double roll)
+	{
+		return CPose3D(.0, .0, .0, yaw, pitch, roll);
+	}
+
+	/** Builds a pose with a translation without rotation \note (New in
+	 * MRPT 2.1.8)
+	 */
+	static CPose3D FromTranslation(double x, double y, double z)
+	{
+		return CPose3D(x, y, z, .0, .0, .0);
+	}
 
 	/** Constructor from a 4x4 homogeneous matrix - the passed matrix can be
 	 * actually of any size larger than or equal 3x4, since only those first
@@ -146,6 +174,15 @@ class CPose3D : public CPose<CPose3D, 6>,
 
 	/** Constructor from a 4x4 homogeneous matrix: */
 	explicit CPose3D(const math::CMatrixDouble44& m);
+
+	/** Builds a pose with a 4x4 homogeneous matrix
+	 * \note (New in MRPT 2.1.8)
+	 */
+	template <class MATRIX>
+	static CPose3D FromHomogeneousMatrix(const MATRIX& m)
+	{
+		return CPose3D(mrpt::math::CMatrixDouble44(m));
+	}
 
 	/** Constructor from a 3x3 rotation matrix and a the translation given as a
 	 * 3-vector, a 3-array, a CPoint3D or a mrpt::math::TPoint3D */
@@ -170,6 +207,16 @@ class CPose3D : public CPose<CPose3D, 6>,
 	{
 	}
 
+	/** Builds a pose with a 3x3 SO(3) rotation matrix and a translation vector.
+	 * \note (New in MRPT 2.1.8)
+	 */
+	template <class MATRIX, class VECTOR>
+	static CPose3D FromRotationAndTranslation(
+		const MATRIX& rot, const VECTOR& t)
+	{
+		return CPose3D(rot, t);
+	}
+
 	/** Constructor from a CPose2D object.
 	 */
 	explicit CPose3D(const CPose2D&);
@@ -192,6 +239,23 @@ class CPose3D : public CPose<CPose3D, 6>,
 
 	/** Constructor from a CPose3DQuat. */
 	explicit CPose3D(const CPose3DQuat&);
+
+	/** Builds a pose from a quaternion (and no translation).
+	 * \note (New in MRPT 2.1.8)
+	 */
+	static CPose3D FromQuaternion(const mrpt::math::CQuaternionDouble& q)
+	{
+		return CPose3D(q, .0, .0, .0);
+	}
+
+	/** Builds a pose from a quaternion and a (x,y,z) translation.
+	 * \note (New in MRPT 2.1.8)
+	 */
+	static CPose3D FromQuaternionAndTranslation(
+		const mrpt::math::CQuaternionDouble& q, double x, double y, double z)
+	{
+		return CPose3D(q, x, y, z);
+	}
 
 	/** Fast constructor that leaves all the data uninitialized - call with
 	 * UNINITIALIZED_POSE as argument */